+++ mocks/include/mocks/BaseApplication.hpp
@@ -20,7 +20,7 @@ class BaseApplication : public EmptyApplication
 {
 public:
     BaseApplication()
-        : settings(this->args, this->settingsDir.path())
+        : settings(this->args, this->settingsDir.path(), /*isTest=*/true)
         , updates(this->paths_, this->settings)
         , theme(this->paths_)
         , fonts(this->settings)
@@ -29,7 +29,7 @@ public:
 
     explicit BaseApplication(const QString &settingsData)
         : EmptyApplication(settingsData)
-        , settings(this->args, this->settingsDir.path())
+        , settings(this->args, this->settingsDir.path(), /*isTest=*/true)
         , updates(this->paths_, this->settings)
         , theme(this->paths_)
         , fonts(this->settings)